+++ pyanalyze/test_signature.py
@@ -1005,6 +1005,38 @@ class TestOverload(TestNameCheckVisitorBase):
             assert_is_value(val3, AnyValue(AnySource.error))
 
     @assert_passes()
+    def test_same_return(self):
+        from pyanalyze.extensions import overload
+        from typing import Any
+        from typing_extensions import Literal
+
+        @overload
+        def overloaded1(x: Any, y: str) -> Literal[2]:
+            pass
+
+        @overload
+        def overloaded1(x: str, y: int) -> Literal[2]:
+            pass
+
+        def overloaded1(x: object, y: object) -> Literal[2]:
+            raise NotImplementedError
+
+        def capybara(x, y):
+            # If multiple overloads match but have the same return type,
+            # don't fall back to Any. This comes up in practice with str.__new__
+            # in typeshed.
+            assert_is_value(overloaded1(x, y), KnownValue(2))
+
+    @assert_passes()
+    def test_nested_class(self):
+        class Outer:
+            class Inner:
+                pass
+
+        def capybara():
+            assert_is_value(Outer.Inner(), TypedValue(Outer.Inner))
+
+    @assert_passes()
     def test_typeshed_overload(self):
         class SupportsWrite:
             def write(self, s: str) -> None: