+++ pyanalyze/type_object.py
@@ -154,7 +154,9 @@ class TypeObject:
     ) -> CanAssign:
         bounds_maps = []
         for member in self.protocol_members:
-            expected = ctx.get_attribute_from_value(self_val, member)
+            expected = ctx.get_attribute_from_value(
+                self_val, member, prefer_typeshed=True
+            )
             # For __call__, we check compatiiblity with the other object itself.
             if member == "__call__":
                 actual = other_val