+++ pyanalyze/typeshed.py
@@ -18,19 +18,21 @@ from .annotations import (
 )
 from .error_code import ErrorCode
 from .extensions import overload, real_overload
-from .safe import is_typing_name
-from .stacked_scopes import uniq_chain
+from .safe import all_of_type, is_typing_name, safe_isinstance
+from .stacked_scopes import uniq_chain, Composite
 from .signature import (
     ConcreteSignature,
     OverloadedSignature,
     SigParameter,
     Signature,
     ParameterKind,
+    make_bound_method,
 )
 from .value import (
     AnySource,
     AnyValue,
     CallableValue,
+    CanAssignContext,
     SubclassValue,
     TypedDictValue,
     TypedValue,
@@ -148,6 +150,7 @@ _TYPING_ALIASES = {
 
 @dataclass
 class TypeshedFinder:
+    ctx: CanAssignContext = field(repr=False)
     verbose: bool = True
     resolver: typeshed_client.Resolver = field(default_factory=typeshed_client.Resolver)
     _assignment_cache: Dict[Tuple[str, ast.AST], Value] = field(
@@ -161,14 +164,20 @@ class TypeshedFinder:
     )
 
     @classmethod
-    def make(cls, options: Options, *, verbose: bool = False) -> "TypeshedFinder":
+    def make(
+        cls,
+        can_assign_ctx: CanAssignContext,
+        options: Options,
+        *,
+        verbose: bool = False,
+    ) -> "TypeshedFinder":
         extra_paths = options.get_value_for(StubPath)
         ctx = typeshed_client.get_search_context()
         ctx = typeshed_client.get_search_context(
             search_path=[*ctx.search_path, *extra_paths]
         )
         resolver = typeshed_client.Resolver(ctx)
-        return TypeshedFinder(verbose, resolver)
+        return TypeshedFinder(can_assign_ctx, verbose, resolver)
 
     def log(self, message: str, obj: object) -> None:
         if not self.verbose:
@@ -176,8 +185,17 @@ class TypeshedFinder:
         print("%s: %r" % (message, obj))
 
     def get_argspec(
-        self, obj: object, *, allow_call: bool = False
+        self,
+        obj: object,
+        *,
+        allow_call: bool = False,
+        type_params: Sequence[Value] = (),
     ) -> Optional[ConcreteSignature]:
+        if isinstance(obj, str):
+            # Synthetic type
+            return self.get_argspec_for_fully_qualified_name(
+                obj, obj, type_params=type_params
+            )
         if inspect.ismethoddescriptor(obj) and hasattr(obj, "__objclass__"):
             objclass = obj.__objclass__
             fq_name = self._get_fq_name(objclass)
@@ -216,16 +234,21 @@ class TypeshedFinder:
         if fq_name is None:
             return None
         return self.get_argspec_for_fully_qualified_name(
-            fq_name, obj, allow_call=allow_call
+            fq_name, obj, allow_call=allow_call, type_params=type_params
         )
 
     def get_argspec_for_fully_qualified_name(
-        self, fq_name: str, obj: object, *, allow_call: bool = False
+        self,
+        fq_name: str,
+        obj: object,
+        *,
+        allow_call: bool = False,
+        type_params: Sequence[Value] = (),
     ) -> Optional[ConcreteSignature]:
         info = self._get_info_for_name(fq_name)
         mod, _ = fq_name.rsplit(".", maxsplit=1)
         sig = self._get_signature_from_info(
-            info, obj, fq_name, mod, allow_call=allow_call
+            info, obj, fq_name, mod, allow_call=allow_call, type_params=type_params
         )
         return sig
 
@@ -242,7 +265,7 @@ class TypeshedFinder:
                 if typ is AbstractSet:
                     return [GenericValue(Collection, (TypeVarValue(T_co),))]
                 if typ is AbstractContextManager or typ is AbstractAsyncContextManager:
-                    return [GenericValue(Generic, (TypeVarValue(T_co),))]
+                    return [GenericValue(Protocol, (TypeVarValue(T_co),))]
                 if typ is Callable or typ is collections.abc.Callable:
                     return None
                 if typ is TypedDict:
@@ -259,7 +282,7 @@ class TypeshedFinder:
                 if fq_name == "collections.abc.Set":
                     return [GenericValue(Collection, (TypeVarValue(T_co),))]
                 elif fq_name == "contextlib.AbstractContextManager":
-                    return [GenericValue(Generic, (TypeVarValue(T_co),))]
+                    return [GenericValue(Protocol, (TypeVarValue(T_co),))]
                 elif fq_name in ("typing.Callable", "collections.abc.Callable"):
                     return None
                 elif is_typing_name(fq_name, "TypedDict"):
@@ -421,32 +444,12 @@ class TypeshedFinder:
                 if info.child_nodes and attr in info.child_nodes:
                     child_info = info.child_nodes[attr]
                     if isinstance(child_info, typeshed_client.NameInfo):
-                        if isinstance(child_info.ast, ast.AnnAssign):
-                            return self._parse_type(
-                                child_info.ast.annotation,
-                                mod,
-                                is_typeddict=is_typeddict,
-                            )
-                        elif isinstance(
-                            child_info.ast, (ast.FunctionDef, ast.AsyncFunctionDef)
-                        ):
-                            decorators = [
-                                self._parse_expr(decorator, mod)
-                                for decorator in child_info.ast.decorator_list
-                            ]
-                            if child_info.ast.returns and decorators == [
-                                KnownValue(property)
-                            ]:
-                                return self._parse_type(child_info.ast.returns, mod)
-                            sig = self._get_signature_from_func_def(
-                                child_info.ast, None, mod, autobind=not on_class
-                            )
-                            if sig is None:
-                                return AnyValue(AnySource.inference)
-                            else:
-                                return CallableValue(sig)
-                        elif isinstance(child_info.ast, ast.Assign):
-                            return UNINITIALIZED_VALUE
+                        return self._get_value_from_child_info(
+                            child_info.ast,
+                            mod,
+                            is_typeddict=is_typeddict,
+                            on_class=on_class,
+                        )
                     assert False, repr(child_info)
                 return UNINITIALIZED_VALUE
             elif isinstance(info.ast, ast.Assign):
@@ -458,6 +461,47 @@ class TypeshedFinder:
             else:
                 return UNINITIALIZED_VALUE
         return UNINITIALIZED_VALUE
+
+    def _get_value_from_child_info(
+        self,
+        node: Union[
+            ast.AST, typeshed_client.OverloadedName, typeshed_client.ImportedName
+        ],
+        mod: str,
+        *,
+        is_typeddict: bool,
+        on_class: bool,
+    ) -> Value:
+        if isinstance(node, ast.AnnAssign):
+            return self._parse_type(node.annotation, mod, is_typeddict=is_typeddict)
+        elif isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef)):
+            decorators = [
+                self._parse_expr(decorator, mod) for decorator in node.decorator_list
+            ]
+            if node.returns and decorators == [KnownValue(property)]:
+                return self._parse_type(node.returns, mod)
+            sig = self._get_signature_from_func_def(
+                node, None, mod, autobind=not on_class
+            )
+            if sig is None:
+                return AnyValue(AnySource.inference)
+            else:
+                return CallableValue(sig)
+        elif isinstance(node, ast.Assign):
+            return UNINITIALIZED_VALUE
+        elif isinstance(node, typeshed_client.OverloadedName):
+            vals = [
+                self._get_value_from_child_info(
+                    subnode, mod, is_typeddict=is_typeddict, on_class=on_class
+                )
+                for subnode in node.definitions
+            ]
+            if all_of_type(vals, CallableValue):
+                sigs = [val.signature for val in vals]
+                if all_of_type(sigs, Signature):
+                    return CallableValue(OverloadedSignature(sigs))
+            return AnyValue(AnySource.inference)
+        assert False, repr(node)
 
     def _get_child_info(
         self, info: typeshed_client.resolver.ResolvedName, attr: str, mod: str
@@ -638,6 +682,7 @@ class TypeshedFinder:
         objclass: Optional[type] = None,
         *,
         allow_call: bool = False,
+        type_params: Sequence[Value] = (),
     ) -> Optional[ConcreteSignature]:
         if isinstance(info, typeshed_client.NameInfo):
             if isinstance(info.ast, (ast.FunctionDef, ast.AsyncFunctionDef)):
@@ -660,6 +705,51 @@ class TypeshedFinder:
                         return None
                     sigs.append(sig)
                 return OverloadedSignature(sigs)
+            elif isinstance(info.ast, ast.ClassDef):
+                new_value, provider = self.get_attribute_recursively(
+                    fq_name, "__new__", on_class=True
+                )
+                sig = None
+                from_init = False
+                if new_value is UNINITIALIZED_VALUE or provider is object:
+                    init_value, provider = self.get_attribute_recursively(
+                        fq_name, "__init__", on_class=True
+                    )
+                    if isinstance(init_value, CallableValue):
+                        sig = init_value.signature
+                        from_init = True
+                elif isinstance(new_value, CallableValue):
+                    sig = new_value.signature
+                if sig is not None:
+                    if safe_isinstance(obj, type):
+                        if allow_call:
+                            if isinstance(sig, Signature):
+                                sig = replace(sig, allow_call=True, callable=obj)
+                            else:
+                                sig = OverloadedSignature(
+                                    [
+                                        replace(sig, allow_call=True, callable=obj)
+                                        for sig in sig.signatures
+                                    ]
+                                )
+                        typ = obj
+                    else:
+                        typ = fq_name
+                    if type_params:
+                        self_val = GenericValue(typ, type_params)
+                    else:
+                        self_val = TypedValue(typ)
+                    if from_init:
+                        sig = sig.replace_return_value(self_val)
+                    else:
+                        self_val = SubclassValue(self_val)
+                    bound_sig = make_bound_method(sig, Composite(self_val))
+                    if bound_sig is None:
+                        return None
+                    sig = bound_sig.get_signature(ctx=self.ctx)
+                    return sig
+
+                return None
             else:
                 self.log("Ignoring unrecognized AST", (fq_name, info))
                 return None