+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Support `__init__` and `__new__` signatures from typeshed (#429)
 - Fix incorrect type inferred for indexing operations on
   subclasses of `list` and `tuple` (#461)
 - Add plugin providing a precise type for `dict.get` calls (#460)